refactor: reorganize POST /g2g-transfer
